requirements.txt
fastapi
sqlalchemy
uvicorn
python-jose
python-dotenv
aiohttp
aiosqlite
passlib[bcrypt]
